refactor(guest-agent): extract cli command building
